+++ tests/UI/ReferrersManager_spec.js
@@ -14,9 +14,9 @@ describe("ReferrersManager", function () {
 
     async function checkUrl(url, page) {
         await page.evaluate(function(){
-            $('[ng-model="urlToCheck"]').val('');
+            $('.url-checker input[type=text]').val('');
         });
-        await page.type('[ng-model="urlToCheck"]', url);
+        await page.type('.url-checker input[type=text]', url);
         await page.click('.checkurlbutton input');
         await page.waitForNetworkIdle();
     }
@@ -28,12 +28,12 @@ describe("ReferrersManager", function () {
     });
 
     it('should identify search engine correct', async function () {
-        await checkUrl('http://www.google.com/xyz/?q=piwik', page);
-        expect(await page.screenshotSelector('[ng-controller="CheckReferrerUrlController"]')).to.matchImage('searchengines_identify');
+        await checkUrl('http://www.google.com/xyz/?q=matomo', page);
+        expect(await page.screenshotSelector('.url-checker')).to.matchImage('searchengines_identify');
     });
 
     it('should open add search engine form', async function () {
-        await (await page.jQuery('[ng-click="showAddSearchEngineForm(true)"]')).click();
+        await page.click('#searchengineTab .add-element');
         await page.waitForTimeout(250);
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('searchengines_add');
     });
@@ -42,18 +42,18 @@ describe("ReferrersManager", function () {
         await page.type("#engineName", 'test');
         await page.type("#engineHost", 'randomhost.com');
         await page.type("#engineParameter", 'q,ghj');
-        await page.click('#searchengineTab [piwik-save-button] input');
+        await page.click('#searchengineTab .matomo-save-button input[value=Save]');
         await page.waitForNetworkIdle();
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('searchengines_list_new');
     });
 
     it('should identify new search engine correct', async function () {
-        await checkUrl('http://randomhost.com/xyz/?ghj=piwik', page);
-        expect(await page.screenshotSelector('[ng-controller="CheckReferrerUrlController"]')).to.matchImage('searchengines_identify_new');
+        await checkUrl('http://randomhost.com/xyz/?ghj=matomo', page);
+        expect(await page.screenshotSelector('.url-checker')).to.matchImage('searchengines_identify_new');
     });
 
     it('should show remove search engine dialog correct', async function () {
-        await page.click('[ng-click="removeEngine(url.url)"]');
+        await page.click('#searchengineTab .delete-element');
         await page.waitForTimeout(500); // wait for animation
         const modal = await page.$('.modal.open');
         expect(await modal.screenshot()).to.matchImage('searchengines_remove_dialog');
@@ -74,22 +74,22 @@ describe("ReferrersManager", function () {
 
     it('should identify social network correct', async function () {
         await checkUrl('http://twitter.com/tweet', page);
-        expect(await page.screenshotSelector('[ng-controller="CheckReferrerUrlController"]')).to.matchImage('social_identify');
+        expect(await page.screenshotSelector('.url-checker')).to.matchImage('social_identify');
     });
 
     it('should disable internal social network list correct', async function () {
-        await (await page.jQuery('[ng-click="setUseDefaultSocials(1)"]')).click();
+        await page.click('#disablesocials');
         await page.waitForNetworkIdle();
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('social_disable');
     });
 
     it('should identify social network correct', async function () {
         await checkUrl('http://twitter.com/tweet', page);
-        expect(await page.screenshotSelector('[ng-controller="CheckReferrerUrlController"]')).to.matchImage('social_identify_empty_list');
+        expect(await page.screenshotSelector('.url-checker')).to.matchImage('social_identify_empty_list');
     });
 
     it('should open add social dialog', async function () {
-        await (await page.jQuery('[ng-click="showAddSocialForm(true)"]')).click();
+        await page.click('#socialTab .add-element');
         await page.waitForTimeout(250);
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('social_add');
     });
@@ -97,24 +97,24 @@ describe("ReferrersManager", function () {
     it('should save new social', async function () {
         await page.type("#socialName", 'test social');
         await page.type("#socialHost", 'randomsocial.com');
-        await page.click('#socialTab [piwik-save-button] input');
+        await page.click('#socialTab .matomo-save-button input[value=Save]');
         await page.waitForNetworkIdle();
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('social_list_new');
     });
 
     it('should identify new social correct', async function () {
         await checkUrl('http://randomsocial.com/123', page);
-        expect(await page.screenshotSelector('[ng-controller="CheckReferrerUrlController"]')).to.matchImage('social_identify_new');
+        expect(await page.screenshotSelector('.url-checker')).to.matchImage('social_identify_new');
     });
 
     it('should enable internal social network list correct', async function () {
-        await (await page.jQuery('[ng-click="setUseDefaultSocials(0)"]')).click();
+        await page.click('#enablesocials');
         await page.waitForNetworkIdle();
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('social_enable');
     });
 
     it('should show remove social network dialog correct', async function () {
-        await page.click('[ng-click="removeSocial(host)"]');
+        await page.click('#socialTab .delete-element');
         await page.waitForTimeout(500); // wait for animation
         const modal = await page.$('.modal.open');
         expect(await modal.screenshot()).to.matchImage('social_remove_dialog');
@@ -123,6 +123,7 @@ describe("ReferrersManager", function () {
     it('should remove new social network correct', async function () {
         await (await page.jQuery(".modal.open a:contains(Yes)")).click();
         await page.waitForNetworkIdle();
+        await page.waitForTimeout(100);
         expect(await page.screenshotSelector('#referrersmanage')).to.matchImage('socials_list');
     });
 });