+++ vue/src/index.ts
@@ -1,0 +1,11 @@
+/*!
+ * Matomo - free/libre analytics platform
+ *
+ * @link https://matomo.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ */
+
+export { default as ManagePage } from './ManagePage/ManagePage.vue';
+export { default as ManageSearchEngines } from './ManageSearchEngines/ManageSearchEngines.vue';
+export { default as ManageSocials } from './ManageSocials/ManageSocials.vue';
+export { default as URLChecker } from './URLChecker/URLChecker.vue';